Become berseker base on token history

+++ src/main/scala/ff/messages/Sensor.scala
@@ -2,7 +2,7 @@ package ff.messages
 
 case class Sensor(
                    raceTrackId: String,
-                   timeStamp: Int,
+                   timeStamp: Long,
                    a: (Int, Int, Int), // acceleration x, y, z
                    g: (Int, Int, Int), // gyro x, y, z
                    m: (Int, Int, Int), // mag x, y, z
